infrastructure/model-serving/requirements-serving.txt
# Requirements for TorchServe model serving

# Core dependencies
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
numpy==1.24.3
scipy==1.11.4

# TorchServe dependencies
torchserve==0.9.0
torch-model-archiver==0.9.0
torch-workflow-archiver==0.2.11

# Monitoring and metrics
prometheus-client==0.19.0
psutil==5.9.6

# Utilities
pyyaml==6.0.1
requests==2.31.0
Pillow==10.1.0

# Model specific
tokenizers==0.15.0
sentencepiece==0.1.99
protobuf==3.20.3

# Performance
onnx==1.15.0
onnxruntime-gpu==1.16.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2